First version of superpower
This first version is for prototyping and testing purposes only. Therefore, the code will be extremely messy and unorganized. Especially the cards.js and the css files.

DB is amazon s3
Server is handled by flask
Frontend is with react

This is a complete working version prototype of the comment drafting tool. The user's data is collected from opted in sources, their instagram posts are collected, and the comments for that post are collected too, sent in bulk to GPT along with the saved user data, and GPT responds with a JSON format containing comment id and the comment response which is sent back to react. It can then be published to instagram by the user. Main limitation is lack of facebook API support for users outside of the business. We can use the app with the company test instagram only. Waiting on facebook business verification for this to be resolved.